FunctionTwitter/requirements.txt
azure-core==1.9.0
azure-storage-blob==12.6.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
idna==2.10
isodate==0.6.0
msrest==0.6.19
oauthlib==3.1.0
pycparser==2.20
PySocks==1.7.1
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.9.0
urllib3==1.26.1
